diffusion: moved types back to the Diffusion.P2P module
