+++ ouroboros-network/src/Ouroboros/Network/Diffusion/NonP2P.hs
@@ -41,8 +41,7 @@ import Ouroboros.Network.Socket (NetworkMutableState, NetworkServerTracers (..),
 import Ouroboros.Network.Context (ExpandedInitiatorContext (..),
            IsBigLedgerPeer (..), MinimalInitiatorContext (..))
 import Ouroboros.Network.ControlMessage (continueForever)
-import Ouroboros.Network.Diffusion.Common hiding (ApplicationsExtra,
-           ArgumentsExtra, TracersExtra, nullTracers)
+import Ouroboros.Network.Diffusion.Common hiding (nullTracers)
 import Ouroboros.Network.ErrorPolicy
 import Ouroboros.Network.IOManager
 import Ouroboros.Network.Mux
